1.8.2

+++ Examples.Unity4/Assets/FairyGUI/Resources/Shaders/FairyGUI-Text.shader
@@ -67,11 +67,11 @@ Shader "FairyGUI/Text"
 					float2 texcoord : TEXCOORD0;
 
 					#ifdef CLIPPED
-					half2 clipPos : TEXCOORD1;
+					float2 clipPos : TEXCOORD1;
 					#endif
 
 					#ifdef SOFT_CLIPPED
-					half2 clipPos : TEXCOORD1;
+					float2 clipPos : TEXCOORD1;
 					#endif
 
 					#ifdef GRAYED
@@ -87,7 +87,7 @@ Shader "FairyGUI/Text"
 
 				#ifdef SOFT_CLIPPED
 				float4 _ClipBox = float4(-2, -2, 0, 0);
-				half4 _ClipSoftness = half4(0, 0, 0, 0);
+				float4 _ClipSoftness = float4(0, 0, 0, 0);
 				#endif
 
 				v2f vert (appdata_t v)
@@ -137,7 +137,7 @@ Shader "FairyGUI/Text"
 					#endif
 
 					#ifdef SOFT_CLIPPED
-					half2 factor = half2(0,0);
+					float2 factor = float2(0,0);
 					if(i.clipPos.x<0)
 						factor.x = (1.0-abs(i.clipPos.x)) * _ClipSoftness.x;
 					else
@@ -150,7 +150,7 @@ Shader "FairyGUI/Text"
 					#endif
 
 					#ifdef CLIPPED
-					half2 factor = abs(i.clipPos);
+					float2 factor = abs(i.clipPos);
 					col.a *= step(max(factor.x, factor.y), 1);
 					#endif
